requirements.txt
# Django
Django==4.1.7
pytest-django
djangorestframework
djangorestframework-simplejwt

# Testing
pytest
pytest-cov
pytest-sugar

# Database
psycopg2

# Code Style
pycodestyle
ruff